danish-tts-trainer/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
phonemizer>=3.2.1
num2words>=0.5.12
espeak-ng-python>=0.1.0
huggingface-hub>=0.16.0
datasets>=2.14.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0
